server/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
filelock==3.17.0
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.56.0
fsspec==2025.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
pillow==11.1.0
pip==24.0
psutil==7.0.0
py-cpuinfo==9.0.0
pymongo==4.8.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.2.73
ultralytics-thop==2.0.14
urllib3==2.3.0
websockets==13.1
Werkzeug==3.1.3